requirements.txt
streamlit==1.28.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.0.0
typing-extensions==4.8.0
uuid==1.30
datetime==5.1
json5==0.9.14
pandas==2.1.0